+++ src/main/java/blusunrize/immersiveengineering/common/data/Recipes.java
@@ -424,6 +424,12 @@ public class Recipes extends RecipeProvider
 				.addInput(IETags.getTagsFor(EnumMetals.ALUMINUM).dust)
 				.addInput(Tags.Items.DYES_YELLOW)
 				.build(out, toRL("blueprint/bullet_flare_yellow"));
+		BlueprintCraftingRecipeBuilder.builder("specialBullet", BulletHandler.getBulletItem(BulletItem.HOMING))
+				.addInput(BulletHandler.emptyCasing)
+				.addInput(Tags.Items.GUNPOWDER)
+				.addInput(new IngredientWithSize(IETags.getTagsFor(EnumMetals.LEAD).nugget, 2))
+				.addInput(Items.ENDER_EYE)
+				.build(out, toRL("blueprint/bullet_homing"));
 
 		BlueprintCraftingRecipeBuilder.builder("electrode", Misc.graphiteElectrode)
 				.addInput(new IngredientWithSize(IETags.hopGraphiteIngot, 4))